fix(go-client): rename clients/go to clients/go-client
The path ending in /go forced users to alias every import since "go"
is a keyword. clients/go-client is unambiguous and follows the same
pattern as the Node and Python client directories.

Also fix publish-npm missing go-client in its needs list.
