+++ .github/workflows/ci.yml
@@ -255,17 +255,17 @@ jobs:
         with:
           go-version: "1.25.5"
           cache-dependency-path: |
-            clients/go/go.sum
+            clients/go-client/go.sum
 
       - name: Install Ginkgo CLI
         run: go install github.com/onsi/ginkgo/v2/ginkgo@v2.28.1
 
       - name: Build Go client
-        working-directory: clients/go
+        working-directory: clients/go-client
         run: go build ./...
 
       - name: Run Go client tests
-        working-directory: clients/go
+        working-directory: clients/go-client
         run: ginkgo ./...
 
   # ─────────────────────────────────────────────────────────────────────────────
