+++ clients/go-client/client.go
