+++ clients/go-client/consumer.go
