+++ clients/go-client/message.go
