+++ clients/go-client/options.go
