+++ clients/go-client/producer.go
