+++ clients/go-client/producer_test.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 
 	pb "github.com/Joessst-Dev/queue-ti/pb"
-	queueti "github.com/Joessst-Dev/queue-ti/clients/go"
+	queueti "github.com/Joessst-Dev/queue-ti/clients/go-client"
 
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
