+++ clients/go-client/suite_test.go
